migration shift_pose_origin_shoulders into shift_pose_origin
